test new engine

+++ Game/include/InputManager.h
@@ -8,5 +8,5 @@ public:
 
     bool IsSpriteClicked(sf::Sprite object, sf::Mouse::Button button, sf::RenderWindow &window);
     bool IsTextClicked(sf::Text text, sf::Mouse::Button button, sf::RenderWindow &window);
-    sf::Vector2i GetMousePosition(sf::RenderWindow &window);
+    sf::Vector2f GetMousePosition(sf::RenderWindow &window);
 };
